bigdata-platform/code/agent-service/requirements.txt
# ====================== LLM ======================
openai>=1.30.0
anthropic>=0.25.0

# ====================== Data ======================
pandas>=2.1.0
numpy>=1.26.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
psycopg2-binary>=2.9.9

# ====================== Visualization ======================
matplotlib>=3.8.0

# ====================== Web ======================
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9

# ====================== Utils ======================
pyyaml>=6.0.1
python-dotenv>=1.0.1
rich>=13.7.0
httpx>=0.27.0

# ====================== Test ======================
pytest>=8.0.0
pytest-cov>=4.1.0
